Move SnapshotCard and DurationSelect to separate dir
