+++ src/app/components/Snapshots/SnapshotCard.tsx
