+++ src/app/pages/ParatimeDashboardPage/Nodes.tsx
@@ -4,7 +4,7 @@ import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import OfflineBoltIcon from '@mui/icons-material/OfflineBolt'
 import InfoIcon from '@mui/icons-material/Info'
-import { SnapshotCard } from './SnapshotCard'
+import { SnapshotCard } from '../../components/Snapshots/SnapshotCard'
 import { COLORS } from '../../../styles/theme/colors'
 import { Layer, useGetRuntimeStatus } from '../../../oasis-indexer/api'
 import { AppErrors } from '../../../types/errors'
