+++ src/app/pages/ParatimeDashboardPage/ParaTimeSnapshot.tsx
@@ -5,7 +5,7 @@ import Typography from '@mui/material/Typography'
 import { useScreenSize } from '../../hooks/useScreensize'
 import { useTheme } from '@mui/material/styles'
 import { styled } from '@mui/material/styles'
-import { DurationSelect } from './DurationSelect'
+import { DurationSelect } from '../../components/Snapshots/DurationSelect'
 import { TransactionsChartCard } from './TransactionsChartCard'
 import { RosePriceCard } from './RosePriceCard'
 import { Nodes } from './Nodes'
