+++ src/app/pages/ParatimeDashboardPage/RosePriceCard.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import Box from '@mui/material/Box'
 import { styled } from '@mui/material/styles'
 import { CoinGeckoReferral } from '../../components/CoinGeckoReferral'
-import { SnapshotCard } from './SnapshotCard'
+import { SnapshotCard } from '../../components/Snapshots/SnapshotCard'
 import { useGetRosePrice } from '../../../coin-gecko/api'
 import { COLORS } from '../../../styles/theme/colors'
 import Typography from '@mui/material/Typography'
